requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
chromadb>=0.4.22
pyyaml>=6.0.1